# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwiftMatch Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applicants:
    get:
      operationId: listApplicants
      security:
        - bearerAuth: []
      tags: [applicants]
      summary: List all applicants (employer subscription or owner only)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ReportForbidden"
        "200":
          description: List of applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Applicant"
    post:
      operationId: createApplicant
      security:
        - bearerAuth: []
      tags: [applicants]
      summary: Create own applicant profile using authenticated account email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicantRequest"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ReportForbidden"
        "409":
          description: Profile already exists
        "201":
          description: Applicant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /applicants/{id}:
    get:
      operationId: getApplicant
      security:
        - bearerAuth: []
      tags: [applicants]
      summary: Get own basic profile free; other profiles require employer access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ReportForbidden"
        "200":
          description: Applicant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplicant
      security:
        - bearerAuth: []
      description: Own profile or owner only. Email identity cannot be changed.
      tags: [applicants]
      summary: Update applicant profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicantRequest"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ReportForbidden"
        "200":
          description: Applicant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing, invalid, expired JWT, missing user, or unconfirmed account
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ReportForbidden:
      description: Subscription required or applicant access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReportAccessError"
  schemas:
...
    ReportAccessError:
      type: object
      required: [error, code, message]
      properties:
        error:
          type: string
          enum: [SUBSCRIPTION_REQUIRED, REPORT_ACCESS_DENIED]
        code:
          type: string
          enum: [SUBSCRIPTION_REQUIRED, REPORT_ACCESS_DENIED]
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    EmploymentHistory:
      type: object
      properties:
        companyName:
          type: string
        position:
          type: string
        yearsStayed:
          type: string
        reasonForLeaving:
          type: string
      required:
        - companyName
        - position
        - yearsStayed
        - reasonForLeaving
    Certificate:
      type: object
      properties:
        name:
          type: string
        issuingOrg:
          type: string
        year:
          type: string
      required:
        - name
        - issuingOrg
        - year
    CharacterReference:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
        email:
          type: string
      required:
        - name
        - relationship
        - phone
        - email
    CreateApplicantRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: ["string", "null"]
        suffix:
          type: ["string", "null"]
        pronoun:
          type: ["string", "null"]
        nickname:
          type: ["string", "null"]
        permanentAddress:
          type: string
        currentAddress:
          type: string
        phoneAreaCode:
          type: string
        phoneNumber:
          type: string
        homePhone:
          type: ["string", "null"]
        email:
          type: string
        skills:
          type: array
          items:
            type: string
          maxItems: 5
        employmentHistory:
          type: array
          items:
            $ref: "#/components/schemas/EmploymentHistory"
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
          maxItems: 3
        references:
          type: array
          items:
            $ref: "#/components/schemas/CharacterReference"
        facebookUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        targetIndustry:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        expectedSalary:
          type: ["string", "null"]
        salaryNegotiable:
          type: boolean
        availabilityDate:
          type: string
        status:
          type: string
          enum: [active, pending, shortlisted, failed, hired]
      required:
        - firstName
        - lastName
        - permanentAddress
        - currentAddress
        - phoneAreaCode
        - phoneNumber
        - email
        - skills
        - employmentHistory
        - certificates
        - references
        - salaryNegotiable
        - availabilityDate
        - status
    UpdateApplicantRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: ["string", "null"]
        suffix:
          type: ["string", "null"]
        pronoun:
          type: ["string", "null"]
        nickname:
          type: ["string", "null"]
        permanentAddress:
          type: string
        currentAddress:
          type: string
        phoneAreaCode:
          type: string
        phoneNumber:
          type: string
        homePhone:
          type: ["string", "null"]
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        employmentHistory:
          type: array
          items:
            $ref: "#/components/schemas/EmploymentHistory"
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        references:
          type: array
          items:
            $ref: "#/components/schemas/CharacterReference"
        facebookUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        targetIndustry:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        expectedSalary:
          type: ["string", "null"]
        salaryNegotiable:
          type: boolean
        availabilityDate:
          type: string
        status:
          type: string
          enum: [active, pending, shortlisted, failed, hired]
    Applicant:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: ["string", "null"]
        suffix:
          type: ["string", "null"]
        pronoun:
          type: ["string", "null"]
        nickname:
          type: ["string", "null"]
        permanentAddress:
          type: string
        currentAddress:
          type: string
        phoneAreaCode:
          type: string
        phoneNumber:
          type: string
        homePhone:
          type: ["string", "null"]
        email:
          type: string
        skills:
          type: array
          items:
            type: string
        employmentHistory:
          type: array
          items:
            $ref: "#/components/schemas/EmploymentHistory"
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        references:
          type: array
          items:
            $ref: "#/components/schemas/CharacterReference"
        facebookUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        targetIndustry:
          type: ["string", "null"]
        targetRole:
          type: ["string", "null"]
        expectedSalary:
          type: ["string", "null"]
        salaryNegotiable:
          type: boolean
        availabilityDate:
          type: string
        status:
          type: string
          enum: [active, pending, shortlisted, failed, hired]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - permanentAddress
        - currentAddress
        - phoneAreaCode
        - phoneNumber
        - email
        - skills
        - employmentHistory
        - certificates
        - references
        - salaryNegotiable
        - availabilityDate
        - status
        - createdAt
        - updatedAt
...