# lib/api-zod/src/generated/types/reportAccessError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwiftMatch Recruitment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportAccessErrorCode } from "./reportAccessErrorCode";
import type { ReportAccessErrorError } from "./reportAccessErrorError";

export interface ReportAccessError {
  error: ReportAccessErrorError;
  code: ReportAccessErrorCode;
  message: string;
}